Fix compile warning.
svn path=/trunk/KDE/kdebase/workspace/; revision=754838

+++ plasma/applets/devicenotifier/itemdelegate.cpp
@@ -75,8 +75,11 @@ void ItemDelegate::paint(QPainter *painter, const QStyleOptionViewItem& option, 
     QFontMetrics subTitleMetrics(subTitleFont);
     QRect textAreaRect = contentRect;
     qreal actualTextWidth = qMax(titleMetrics.width(titleText), subTitleMetrics.width(subTitleText));
-    textAreaRect.adjust(decorationRect.width() + ICON_TEXT_MARGIN - 3, 0, -(textRect.width() - actualTextWidth) + 3, 1);
-        
+    textAreaRect.adjust(decorationRect.width() + ICON_TEXT_MARGIN - 3,
+                        0,
+                        (int)(-(textRect.width() - actualTextWidth) + 3),
+                        1);
+
     if (hover) {
         painter->save();
         painter->setPen(Qt::NoPen);
